Make new gcc -Wall happy

+++ Parser/grammar.c
@@ -217,7 +217,7 @@ translabel(g, lb)
 	}
 	
 	if (lb->lb_type == STRING) {
-		if (isalpha(lb->lb_str[1]) || lb->lb_str[1] == '_') {
+		if (isalpha((int)(lb->lb_str[1])) || lb->lb_str[1] == '_') {
 			char *p;
 			if (Py_DebugFlag)
 				printf("Label %s is a keyword\n", lb->lb_str);
